\name{kjhutils-package}
\alias{kjhutils-package}
\alias{kjhutils}
\docType{package}
\title{
  Convenience functions for data analysis. 
}
\description{
  Strictly local package of useful functions for data analysis.
}
\details{
\tabular{ll}{
Package: \tab kjhutils\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-05-04\cr
License: \tab GPL-2\cr
}
}
\author{
  Kieran Healy.
  
  Kieran Healy: Who to complain to <kjhealy@gmail.com>

}
\references{

}

\keyword{ package }
\seealso{
}
\examples{

}
