\name{makeFootnote}
\alias{makeFootnote}
\title{makeFootnote}
\usage{
  makeFootnote(footnoteText = format(Sys.time(), "\%d \%b \%Y"),
    size = 0.7, color = grey(0.5))
}
\arguments{
  \item{footnoteText}{Defaults to date stamp.}

  \item{size}{}

  \item{color}{}
}
\value{
  plot footnote
}
\description{
  Footnote a plot.
}
\author{
  Kieran Healy
}

