% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{ptable}
\alias{ptable}
\title{ptable}
\usage{
ptable(..., by.pct = 1, dig = 2)
}
\arguments{
\item{...}{Like table, the vectors to be tablulated.}

\item{by.pct}{Which dimension to calculate the perentages on. 1 is row, 2 is column, etc.}

\item{dig}{Number of digits after the decimal point in the output}
}
\value{
A prop.table object
}
\description{
n-way contingency tables showing row/column percents. Nicer than raw frequencies.
convenience function
}
\author{
Kieran Healy
}
