\name{vplayout}
\alias{vplayout}
\title{vplayout}
\usage{
  vplayout(x, y)
}
\arguments{
  \item{x}{n rows in viewport}

  \item{y}{n columns in viewport}
}
\value{
  A viewport for plotting with x rows and y columns
}
\description{
  Refer to viewport layouts easily (Wickham)
}
\author{
  Kieran Healy
}

