\name{ptable}
\alias{ptable}
\title{ptable}
\usage{
  ptable(..., by.pct = 1, dig = 2)
}
\arguments{
  \item{...}{Like table, the vectors to be tablulated.}

  \item{by.pct}{Which dimension to calculate the
  perentatges}

  \item{dig}{Number of digits after the decimal point in
  the output}
}
\value{
  A nice table
}
\description{
  n-way contingency tables showing row/column/etc percents
  instead
}
\author{
  Kieran Healy
}

